refactor: group catalog authoring workflows
