+++ apps/backend/src/catalog/authoring/versions/index.ts
@@ -1,13 +1,13 @@
-import { unsafeTransaction } from "../../database/core";
+import { unsafeTransaction } from "../../../database/core";
 import type {
   CatalogPackageVersion,
   CreateCatalogPackageVersionFromWorkspaceInput,
   CreateCatalogPackageVersionInput,
   UpdateCatalogPackageVersionStatusInput,
-} from "../types";
+} from "../../types";
 import {
   createCatalogPackageVersionFromCardsInExecutor,
-} from "./versionCreation";
+} from "./creation";
 import {
   createCatalogPackageVersionFromWorkspaceSelectionInExecutor,
 } from "./workspaceSnapshots";
@@ -19,7 +19,7 @@ import {
 
 export {
   createCatalogPackageVersionFromCardsInExecutor,
-} from "./versionCreation";
+} from "./creation";
 export {
   createCatalogPackageVersionFromWorkspaceSelectionInExecutor,
 } from "./workspaceSnapshots";
